+++ client/client/src/adminComponents/features/CreateProduct/roots/ProductManager.jsx
@@ -1,14 +1,11 @@
 import React from "react";
 import styles from "./ProductManager.module.css";
 import CreateProducts from "./CreateProducts";
-import ProductList from "../components/ProductList";
+
 
 const ProductManager = () => {
   return (
     <div className={styles.container}>
-      <div className={styles.childList}>
-        <ProductList />
-      </div>
       <div className={styles.childCard}>
         <CreateProducts/>
       </div>
