+++ client/client/src/adminComponents/features/ProductList/roots/ProductsManager.module.css
@@ -1,0 +1,11 @@
+.container {
+  display: flex;
+  justify-content: space-around;
+}
+
+
+.childList {
+  text-align: center;
+  min-width: 500px;
+  width: auto;
+}
