+++ client/client/src/redux/actions/action.js
@@ -43,7 +43,7 @@ import {
 } from "../actions-types/actions-types";
 import axios from "axios";
 
-const SERVER_URL = "https://surf-4i7c.onrender.com/surf";
+const SERVER_URL = "http://localhost:3001/surf";
 
 //PRODUCTS ACTIONS
 export const getAllProducts = () => {
